Set statement_timeout for DB connections
The patch sets the statement_timeout parameter for the ocitysmap
sessions to 18mn. This should be enough. I checked that an exception
is raised when the timeout appears (=> ocitysmap crashes
immediatly). The other sessions (eg. osm2pgsql) should not be
affected. Issue "show statement_timeout" on a new DB connection to
make sure (should reply '0').

This is a "quick and dirty fix": the timeout should be configurable
from a config file or from the cmdline. For now it's hard-coded in
street_index.py.

Patch by David Decotigny <d2@maposmatic.org>