Use psycopg2 instead of pygresql (aka. pgdb)
Django relies on psycopg2 for its postgres backend, whereas ocitysmap
was using another (pgdb). This patch addresses task #9666: it removes
the pygresql dependency and changes it for psycopg2.

The major part of the patch deals with unicode issues, because
psycopg2 can handle unicode strings directly. Actual;ly, that's what
it does when it is being used along django. So we force this same
behavior as well when we run ocitysmap in plain "command line" mode.