Implement support for RTL languages in street index
We introduce a isrtl() method in the i18n classes. By default, this
method returns False (for non-RTL languages), but RTL languages can
return True to enable the RTL rendering mode for the map and the
index.

For now, the RTL mode :

 * Changes the rendering of the index columns: the first column is on
   the right.

 * Changes on index entries are rendered: the street name is on the
   right while the location on the map squares is on the left.

We must ask Arabic person to know whether the square on the map should
also be rendered from right to left (i.e column A on the right).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>