Foundation for the new OCitySMap rendering module
This new OCitySMap version aims at providing a more print-oriented
map rendering engine.

Its main features include being aware of paper sizes and constraints,
but also providing different "renderers" implementing rendering layouts
for the maps. For now, only the PlainRenderer which renderers a
full-page map is implemented, but renderers that also generate an index
will be available.

The main idea is that we no longer generate a separate index. The
generated map includes (or not) an index depending on the rendering
layout used to render it.

Finally, each renderer is able to tell which paper sizes are compatible
with the bounding box from a "geographic" resolution.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>