Continue heavy work on OCitySMap 2
Re-implement configuration file parsing to get access to the database.
The request timeout is now configurable in the 'datasource'
configuration section as 'request_timeout'. It defaults to 15 minutes.

Re-integrate the i18n and languages support.

Re-implement the map grid support. From the corrected geographical
bounding box, the Grid object can generate an ESRI Shapefile with the
grid lines. The grid size automatically adjusts to lower sizes if the
bounding box is smaller.

Stub implementation of the StreetIndex object that will handle data
gathering, outputting as CSV and rendering into a given surface.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>