Fixed layout for RTL languages
With RTL languages, ALIGN_RIGHT is interpreted as ALIGN_LEFT by
default in the layout objects, but only when the text is "strongly"
RTL (eg. arabic only, no western text intermixed). This breaks the
asumptions we made for our layout of the index streets. This patch
tells the pango layout manager to ignore this and always interpret
ALIGN_RIGHT as ALIGN_RIGHT.

Signed-off-by: David Decotigny <d2@maposmatic.org>