Provide a unified way to get bounding box and contour from OSM IDs
The new get_geographic_info method in OCitySMap allows one to query the
PostGIS database for the commonly-needed geographic information about a
set of OSM IDs.

This is used by OCitySMap itself, of course, for map rendering when it
needs to find the bounding box and contour of the requested OSM ID. This
will also be used by the Nominatim search gateway in MapOSMatic in the
near future.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>