Map street paths onto grid, pass 1
To map the street paths onto a Grid, we proceed in 2 passes since the
grid is not known before the streets are determined (size of the
rendered bbox depends on the number of streets => grid position over
the city's rendering is not fixed). This patch allows to optimize the
number of GIS queries: we only need to perform them in pass 1.

First pass gets the positions of the most distant points from the
street. This is what this patch implements. The second pass will then
map these points onto the grid.

This patch also changes the IndexItem structure to include the 2 most
distant endpoints of the item.

We only deal with the streets for now. The code for amenities is dummy
at that point; we'll work on this later.

Signed-off-by: David Decotigny <d2@maposmatic.org>