ocitysmap/support: add support for storing the age of the OSM data
This commit creates a new maposmatic_admin table, which contains a
single column and a single entry, used to store the age of the OSM
data that is currently inside the GIS database. This table is located
on the GIS database side (and *NOT* on the Django database side).

This table must be created once on the GIS database by running the
setup-maposmatic-admin.sql.

This table is then updated by the planet-update.sh script every time
it completes an update of the OSM data. System administrators of the
production and development servers are therefore now requested to use
the planet-update.sh script for the OcitySMap repository on the
servers, and to replicate on the Git repository any change done on
those scripts.

The contents of this table will be used by later commits to show the
replication lag on the MapOSMatic.org website and the date of the OSM
data on the generated maps.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>