ocitysmap2: show OSM data last update timestamp on rendered maps
We use the maposmatic_admin table that is now part of the GIS database
and updated by the planet-update.sh script to show the timestamp of
the OSM data on rendered maps. If this table does not exist, we
gracefully fallback to an "unknown" timestamp, so that the rendering
doesn't fail.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>