coords: add BoundingBox::as_javascript() method
This method returns a string containing a bit of Javascript code that
is useful to render the Bounding Box with the Leaflet Javascript
code. It allows to quickly have a graphical representation of multiple
bounding boxes on a slippy map. This is very useful to understand if
the computations done on bounding boxes are correct.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>