indexlib: do not raise exception when index is empty
The StreetIndex constructor was raising an exception when no street
was found. With the multi-page renderer, it is very possible to have
pages of map without any street or other elements to put in the index,
but for convenience we still want to get a StreetIndex object, even if
its ._categories list is empty.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>