support: add bounding-box-debug.html
This short HTML page is helpful to debug bounding box related
problems, by visualizing a set of bounding boxes on a slippy map. Many
bounding boxes can be added using the BoundingBox() function.

Note that the OcitySMap BoundingBox object has a .as_javascript()
method that directly generates a string that is suitable for
copy/paste into this file.

This file has been _very_ helpful to debug some initial problems of
the multi-page renderer.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>