Shorter German translations in ocitysmap.po
To address bug #36166, the translation for Public Building(s) is
replaced with the more literal 'Gebäude(n)', which is slightly
shorter.  Similarly for Education, which was translated as
'Educational institutions', it is replaced this with a more literal
(and shorter) Education -> Bildung.

Of course it doesn't fix the bug, as such, but may help create a
better index on German maps until the bug is fixed properly.

Signed-off-by: Jeroen van Rijn <jvrnix@gmail.com>