i18n: reorder spanish determinants
Due to the order of spanish determinants, a street such as "Calle de
los Huertos" was indexed under "los Huertos (Calle de)" instead of
"Huertos (Calle de los)". Same problem for "de la" or "de las". By
reordering the determinants so that the most specific ones are
considered first, we can achieve the desired behaviour.

See
http://www.maposmatic.org/rendered//046017_2012-04-27_11-51_MirafloresdelaSierraSpain.png
for an example of many current problems.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>